pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "exonware-xwsystem"
dynamic = ["version"]
description = "Enterprise-grade Python framework with AI-powered performance optimization, 24 serialization formats, military-grade security, automatic memory leak prevention, circuit breakers, and production monitoring - replaces 50+ dependencies"
readme = "README.md"
requires-python = ">=3.8"
keywords = [ "enterprise", "framework", "serialization", "security", "performance", "monitoring", "threading", "memory-management", "circuit-breaker", "object-pool", "crypto", "ai-optimization", "json", "yaml", "msgpack", "bson", "cbor", "exonware",]
classifiers = [ "Development Status :: 4 - Beta", "Intended Audience :: Developers", "Intended Audience :: Information Technology", "Intended Audience :: System Administrators", "License :: OSI Approved :: MIT License", "Operating System :: OS Independent", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Topic :: Software Development :: Libraries :: Python Modules", "Topic :: System :: Systems Administration", "Topic :: Security :: Cryptography", "Topic :: Internet :: WWW/HTTP", "Topic :: Database", "Topic :: System :: Monitoring",]
dependencies = [
    # Core only - minimal dependencies for basic functionality
    "typing-extensions>=4.0.0",
]
[[project.authors]]
name = "Eng. Muhammad AlShehri"
email = "connect@exonware.com"

[project.license]
text = "MIT"

[project.optional-dependencies]
lazy = [
    # Marker package that enables lazy functionality
    "exonware-xwlazy>=0.1.0",
]
full = [
    # Core serialization formats (actually used in codebase)
    "PyYAML>=5.4.0",
    "pyyaml-include>=1.3.0",
    "ruamel.yaml>=0.17.0",
    "tomli-w>=1.0.0",
    "tomli>=2.0.0",
    "dicttoxml>=1.7.0",
    "xmltodict>=0.12.0",
    "defusedxml>=0.7.0",
    "lxml>=4.9.0",
    "pymongo>=3.12.0",
    "msgpack>=1.0.0",
    "cbor2>=5.0.0",
    "json5>=0.9.0",
    
    # Core framework dependencies
    "psutil>=5.8.0",
    "httpx>=0.24.0",
    "cryptography>=3.4.0",
    "bcrypt>=4.0.0",
    "aiofiles>=0.8.0",
    "PyJWT>=2.6.0",
    "colorama>=0.4.0",
    "requests>=2.28.0",
    "pytz>=2023.3",
    "orjson>=3.8.0",
    "ijson>=3.2.0",
    "jsonpointer>=2.0.0",
    "jsonpatch>=1.33.0",
    "jsonpath-ng>=1.6.0",
    "jsonschema>=4.17.0",
    "rpds-py>=0.10.0",
    "xxhash>=3.2.0",
    "msgspec>=0.11.0",
    "rtoml>=0.7.0",
    
    # Monitoring & observability
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    
    # XML validation
    "xmlschema>=2.0.0",
]

# NOTE: Enterprise formats (~87 MB) moved to exonware-xwformats
# Install with: pip install exonware-xwformats
# This adds 16 formats:
# - Schema (7): Protobuf, Avro, Parquet, Thrift, ORC, Cap'n Proto, FlatBuffers
# - Scientific (5): HDF5, Feather, Zarr, NetCDF, MAT
# - Database (3): LMDB, GraphDB, LevelDB
# - Binary (1): UBJSON

[project.urls]
Homepage = "https://exonware.com"
Repository = "https://github.com/exonware/xwsystem"
Documentation = "https://github.com/exonware/xwsystem#readme"

[tool.black]
line-length = 88
target-version = [ "py38",]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "0.0.1.408"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = [ "tests",]
python_files = [ "test_*.py",]
python_classes = [ "Test*",]
python_functions = [ "test_*",]
addopts = "-v --tb=short"
markers = [ "xwsystem_core: Core functionality tests", "xwsystem_unit: Unit tests for individual components", "xwsystem_integration: Integration tests between components", "xwsystem_config: Configuration and setup tests", "xwsystem_performance: Performance and benchmarking tests", "xwsystem_security: Security validation tests", "xwsystem_threading: Threading and concurrency tests", "xwsystem_io: I/O operations tests", "xwsystem_structures: Data structure tests", "xwsystem_patterns: Design pattern tests", "xwsystem_monitoring: Monitoring and metrics tests",]

[tool.hatch.version]
path = "src/exonware/xwsystem/version.py"

[tool.hatch.build.targets.wheel]
packages = [ "src/exonware",]
py-modules = [ "src/xwsystem",]

[tool.hatch.build.targets.sdist]
include = [ "/src", "/tests", "/README.md", "/.github",]
